refactor: reorganize endpoints into appropriate set of modules
This commit moves various endpoints into more appropriate python
modules. It is a refactor only, with no functional impact.

The initial motivation was to ensure the publish APIs and service
APIs don't live in the same module. These endpoints are quite
unrelated and it doesn't make sense to me that their implementation,
and associated tests, were all dumped into the same 'gateway'
module.

Going a bit further, it seems reasonable to have the python modules
matching the OpenAPI tags used for the endpoints. Both the tagging
and the split between modules are ways of grouping related
functionality and there's no reason to have one grouping in the
source code layout and a different grouping in the API spec, so let's
rename the modules to make them match. This also allows applying
default tags in the APIRouter of each module, so you can't forget
to apply them when adding new endpoints.
