+++ exodus_gw/routers/publish.py
@@ -3,41 +3,18 @@ from os.path import basename
 from typing import List, Union
 from uuid import UUID
 
-from fastapi import APIRouter, Depends, Request
+from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
 
-from .. import models, schemas, worker
-from ..auth import CallContext, call_context
+from .. import models, schemas
 from ..aws.dynamodb import write_batches
 from ..crud import create_publish, get_publish_by_id, update_publish
+from ..database import get_db
 from ..settings import get_environment, get_settings
 
 LOG = logging.getLogger("exodus-gw")
 
-router = APIRouter()
-
-
-def get_db(request: Request):
-    return request.state.db
-
-
-@router.get("/healthcheck", tags=["service"])
-def healthcheck():
-    """Returns a successful response if the service is running."""
-    return {"detail": "exodus-gw is running"}
-
-
-@router.get("/healthcheck-worker", tags=["service"])
-def healthcheck_worker():
-    """Returns a successful response if background workers are running."""
-
-    msg = worker.ping.send()
-
-    # If we don't get a response in time, this will raise an exception and we'll
-    # respond with a 500 error, which seems reasonable.
-    result = msg.get_result(block=True, timeout=5000)
-
-    return {"detail": "background worker is running: ping => %s" % result}
+router = APIRouter(tags=["publish"])
 
 
 @router.post(
@@ -58,7 +35,6 @@ async def publish(env: str, db: Session = Depends(get_db)) -> models.Publish:
 @router.put(
     "/{env}/publish/{publish_id}",
     status_code=200,
-    tags=["publish"],
 )
 async def update_publish_items(
     env: str,
@@ -79,7 +55,6 @@ async def update_publish_items(
 @router.post(
     "/{env}/publish/{publish_id}/commit",
     status_code=200,
-    tags=["publish"],
 )
 async def commit_publish(
     env: str, publish_id: UUID, db: Session = Depends(get_db)
@@ -117,40 +92,3 @@ async def commit_publish(
             await write_batches(env, items + last_items, delete=True)
 
     return {}
-
-
-@router.get(
-    "/whoami",
-    response_model=CallContext,
-    tags=["service"],
-    responses={
-        200: {
-            "description": "returns caller's auth context",
-            "content": {
-                "application/json": {
-                    "example": {
-                        "client": {
-                            "roles": ["someRole", "anotherRole"],
-                            "authenticated": True,
-                            "serviceAccountId": "clientappname",
-                        },
-                        "user": {
-                            "roles": ["viewer"],
-                            "authenticated": True,
-                            "internalUsername": "someuser",
-                        },
-                    }
-                }
-            },
-        }
-    },
-)
-def whoami(context: CallContext = Depends(call_context)):
-    """Return basic information on the caller's authentication & authorization context.
-
-    This endpoint may be used to determine whether the caller is authenticated to
-    the exodus-gw service, and if so, which set of role(s) are held by the caller.
-
-    It is a read-only endpoint intended for diagnosing authentication issues.
-    """
-    return context
