+++ tests/routers/test_publish.py
@@ -2,21 +2,7 @@ import mock
 import pytest
 from fastapi import HTTPException
 
-from exodus_gw import models
-from exodus_gw.routers import gateway
-
-
-def test_healthcheck():
-    assert gateway.healthcheck() == {"detail": "exodus-gw is running"}
-
-
-def test_healthcheck_worker(stub_worker):
-    # This test exercises "real" dramatiq message handling via stub
-    # broker & worker, demonstrating that messages can be used from
-    # within tests.
-    assert gateway.healthcheck_worker() == {
-        "detail": "background worker is running: ping => pong"
-    }
+from exodus_gw import models, routers
 
 
 @pytest.mark.asyncio
@@ -29,7 +15,7 @@ def test_healthcheck_worker(stub_worker):
     ],
 )
 async def test_publish_env_exists(env, mock_db_session):
-    publish = await gateway.publish(env=env, db=mock_db_session)
+    publish = await routers.publish.publish(env=env, db=mock_db_session)
     assert isinstance(publish, models.Publish)
 
 
@@ -37,7 +23,7 @@ async def test_publish_env_exists(env, mock_db_session):
 async def test_publish_env_doesnt_exist(mock_db_session):
     env = "foo"
     with pytest.raises(HTTPException) as exc_info:
-        await gateway.publish(env=env, db=mock_db_session)
+        await routers.publish.publish(env=env, db=mock_db_session)
     assert exc_info.value.status_code == 404
     assert exc_info.value.detail == "Invalid environment='foo'"
 
@@ -59,7 +45,7 @@ async def test_update_publish_items_env_exists(
     items = mock_item_list[0] if env == "test3" else mock_item_list
 
     assert (
-        await gateway.update_publish_items(
+        await routers.publish.update_publish_items(
             env=env,
             publish_id=publish_id,
             items=items,
@@ -77,7 +63,7 @@ async def test_update_publish_items_env_doesnt_exist(
     publish_id = "123e4567-e89b-12d3-a456-426614174000"
 
     with pytest.raises(HTTPException) as exc_info:
-        await gateway.update_publish_items(
+        await routers.publish.update_publish_items(
             env=env,
             publish_id=publish_id,
             items=mock_item_list,
@@ -97,8 +83,8 @@ async def test_update_publish_items_env_doesnt_exist(
         "test3",
     ],
 )
-@mock.patch("exodus_gw.routers.gateway.write_batches")
-@mock.patch("exodus_gw.routers.gateway.get_publish_by_id")
+@mock.patch("exodus_gw.routers.publish.write_batches")
+@mock.patch("exodus_gw.routers.publish.get_publish_by_id")
 async def test_commit_publish(
     mock_get_publish,
     mock_write_batches,
@@ -110,7 +96,7 @@ async def test_commit_publish(
     mock_write_batches.return_value = True
 
     assert (
-        await gateway.commit_publish(
+        await routers.publish.commit_publish(
             env=env, publish_id=mock_publish.id, db=mock_db_session
         )
         == {}
@@ -126,12 +112,12 @@ async def test_commit_publish(
 
 
 @pytest.mark.asyncio
-@mock.patch("exodus_gw.routers.gateway.get_publish_by_id")
+@mock.patch("exodus_gw.routers.publish.get_publish_by_id")
 async def test_commit_publish_env_doesnt_exist(mock_publish, mock_db_session):
     env = "foo"
 
     with pytest.raises(HTTPException) as exc_info:
-        await gateway.commit_publish(
+        await routers.publish.commit_publish(
             env=env, publish_id=mock_publish.id, db=mock_db_session
         )
 
@@ -140,15 +126,15 @@ async def test_commit_publish_env_doesnt_exist(mock_publish, mock_db_session):
 
 
 @pytest.mark.asyncio
-@mock.patch("exodus_gw.routers.gateway.write_batches")
-@mock.patch("exodus_gw.routers.gateway.get_publish_by_id")
+@mock.patch("exodus_gw.routers.publish.write_batches")
+@mock.patch("exodus_gw.routers.publish.get_publish_by_id")
 async def test_commit_publish_write_failed(
     mock_get_publish, mock_write_batches, mock_publish, mock_db_session
 ):
     mock_get_publish.return_value = mock_publish
     mock_write_batches.side_effect = [False, True]
 
-    await gateway.commit_publish(
+    await routers.publish.commit_publish(
         env="test", publish_id=mock_publish.id, db=mock_db_session
     )
 
@@ -162,15 +148,15 @@ async def test_commit_publish_write_failed(
 
 
 @pytest.mark.asyncio
-@mock.patch("exodus_gw.routers.gateway.write_batches")
-@mock.patch("exodus_gw.routers.gateway.get_publish_by_id")
+@mock.patch("exodus_gw.routers.publish.write_batches")
+@mock.patch("exodus_gw.routers.publish.get_publish_by_id")
 async def test_commit_publish_entry_point_files_failed(
     mock_get_publish, mock_write_batches, mock_publish, mock_db_session
 ):
     mock_get_publish.return_value = mock_publish
     mock_write_batches.side_effect = [True, False, True]
 
-    await gateway.commit_publish(
+    await routers.publish.commit_publish(
         env="test", publish_id=mock_publish.id, db=mock_db_session
     )
 
@@ -182,10 +168,3 @@ async def test_commit_publish_entry_point_files_failed(
         ],
         any_order=False,
     )
-
-
-def test_whoami():
-    # All work is done by fastapi deserialization, so this doesn't actually
-    # do anything except return the passed object.
-    context = object()
-    assert gateway.whoami(context=context) is context
