+++ tests/routers/test_service.py
@@ -1,0 +1,26 @@
+import mock
+import pytest
+from fastapi import HTTPException
+
+from exodus_gw import models
+from exodus_gw.routers import service
+
+
+def test_healthcheck():
+    assert service.healthcheck() == {"detail": "exodus-gw is running"}
+
+
+def test_healthcheck_worker(stub_worker):
+    # This test exercises "real" dramatiq message handling via stub
+    # broker & worker, demonstrating that messages can be used from
+    # within tests.
+    assert service.healthcheck_worker() == {
+        "detail": "background worker is running: ping => pong"
+    }
+
+
+def test_whoami():
+    # All work is done by fastapi deserialization, so this doesn't actually
+    # do anything except return the passed object.
+    context = object()
+    assert service.whoami(context=context) is context
