+++ tests/routers/upload/test_manage_mpu.py
@@ -5,7 +5,7 @@ import pytest
 from fastapi import HTTPException
 
 from exodus_gw.aws.util import xml_response
-from exodus_gw.routers.s3 import abort_multipart_upload, multipart_upload
+from exodus_gw.routers.upload import abort_multipart_upload, multipart_upload
 
 TEST_KEY = "b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c"
